[Improvement-3369][api] Introduce monitor, processinstance and queue service interface for clear code (#4765)
* [Improvement-3369][api] Introduce monitor, processinstance and queue service interface for clear code
